embedding_service/pyproject.toml
[project]
name = "twodirect-embedding-service"
version = "0.1.0"
description = "Multimodal embedding service for twodirect product search using Jina CLIP v2"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "transformers>=4.47.0",
    "torch>=2.5.1",
    "torchvision>=0.20.0",
    "timm>=1.0.0",
    "einops>=0.8.0",
    "pillow>=11.0.0",
    "numpy>=2.2.0",
    "supabase>=2.11.0",
    "psycopg2-binary>=2.9.10",
    "pgvector>=0.3.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
production = [
    "gunicorn>=23.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = []

